change quality type to int

+++ src/main/java/ru/unisuite/imageresizer/ImageResizer.java
@@ -6,11 +6,11 @@ import java.io.OutputStream;
 
 public interface ImageResizer {
 	
-	public void resizeByWidth(InputStream is, int width, OutputStream os, int length, float quality) throws IOException;
+	public void resizeByWidth(InputStream is, int width, OutputStream os, int length, Integer quality) throws IOException;
 	
-	public void resize(InputStream is, int width, int height, OutputStream os, float quality) throws IOException;
+	public void resize(InputStream is, int width, int height, OutputStream os, Integer quality) throws IOException;
 	
-	public void resizeByHeight(InputStream is, int height, OutputStream os, float quality) throws IOException;
+	public void resizeByHeight(InputStream is, int height, OutputStream os, Integer quality) throws IOException;
